requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.67.0
anyio==4.10.0
asttokens==3.0.0
asyncer==0.0.8
attrs==25.3.0
audioop-lts==0.2.2
autopep8==2.3.2
backoff==2.2.1
bcrypt==4.3.0
bidict==0.23.1
black==25.1.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
chainlit==2.8.0
charset-normalizer==3.4.3
chevron==0.14.0
chromadb==1.0.21
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
cuid==0.4
dataclasses-json==0.6.7
debugpy==1.8.16
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
durationpy==0.10
executing==2.2.1
fastapi==0.116.1
filelock==3.19.1
filetype==1.2.0
flake8==7.3.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.41
google-auth==2.40.3
googleapis-common-protos==1.70.0
griffe==1.14.0
grpcio==1.75.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.6
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kubernetes==33.1.0
Lazify==0.4.0
literalai==0.1.201
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mcp==1.14.0
mdurl==0.1.2
mmh3==5.2.0
monotonic==1.6
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
nbqa==1.9.1
nest-asyncio==1.6.0
numpy==2.3.3
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.107.3
openai-agents==0.3.0
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-alephalpha==0.47.1
opentelemetry-instrumentation-anthropic==0.47.1
opentelemetry-instrumentation-bedrock==0.47.1
opentelemetry-instrumentation-chromadb==0.47.1
opentelemetry-instrumentation-cohere==0.47.1
opentelemetry-instrumentation-crewai==0.47.1
opentelemetry-instrumentation-google-generativeai==0.47.1
opentelemetry-instrumentation-groq==0.47.1
opentelemetry-instrumentation-haystack==0.47.1
opentelemetry-instrumentation-lancedb==0.47.1
opentelemetry-instrumentation-langchain==0.47.1
opentelemetry-instrumentation-llamaindex==0.47.1
opentelemetry-instrumentation-logging==0.58b0
opentelemetry-instrumentation-marqo==0.47.1
opentelemetry-instrumentation-mcp==0.47.1
opentelemetry-instrumentation-milvus==0.47.1
opentelemetry-instrumentation-mistralai==0.47.1
opentelemetry-instrumentation-ollama==0.47.1
opentelemetry-instrumentation-openai==0.47.1
opentelemetry-instrumentation-openai-agents==0.47.1
opentelemetry-instrumentation-pinecone==0.47.1
opentelemetry-instrumentation-qdrant==0.47.1
opentelemetry-instrumentation-redis==0.58b0
opentelemetry-instrumentation-replicate==0.47.1
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-sagemaker==0.47.1
opentelemetry-instrumentation-sqlalchemy==0.58b0
opentelemetry-instrumentation-threading==0.58b0
opentelemetry-instrumentation-together==0.47.1
opentelemetry-instrumentation-transformers==0.47.1
opentelemetry-instrumentation-urllib3==0.58b0
opentelemetry-instrumentation-vertexai==0.47.1
opentelemetry-instrumentation-watsonx==0.47.1
opentelemetry-instrumentation-weaviate==0.47.1
opentelemetry-instrumentation-writer==0.47.1
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.58b0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==2.3.2
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.4.0
posthog==3.25.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycodestyle==2.14.0
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
setuptools==75.6.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.47.3
sympy==1.14.0
syncer==2.0.3
tenacity==9.1.2
tokenize_rt==6.2.0
tokenizers==0.22.0
tomli==2.2.1
tornado==6.5.2
tqdm==4.67.1
traceloop-sdk==0.47.1
traitlets==5.14.3
typer==0.17.4
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.3
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
